+++ lib/oxide_api/timeseries.ex
@@ -5,8 +5,8 @@ defmodule OxideApi.Timeseries do
 
   alias OxideApi.Client
 
-  @spec query(Client.t(), map()) :: Client.result()
-  def query(%Client{} = client, body) when is_map(body) do
-    Client.post(client, "/v1/timeseries/query", body)
+  @spec query(Client.t(), map(), keyword()) :: Client.result()
+  def query(%Client{} = client, body, params \\ []) when is_map(body) do
+    Client.post(client, "/v1/timeseries/query", body, params: params)
   end
 end